Use gates to threshold matrices, use just 7 markers
In our original tutorial we only looked at combinations of 7 markers.

I had erroneously set all negative values in the matrices to 0, when it
should have been only cells that were negative for that marker by gating
that are set to 0. See https://github.com/RGLab/COMPASS/issues/81 for
context.